moved all sql related code to a seperate class. Fixes #11
