+++ backend/server.py
@@ -1,61 +1,23 @@
 from flask import Flask, request, jsonify
 from flask_login import LoginManager, login_user, current_user, login_required
-import mysql.connector
 import os
 from dotenv import load_dotenv
-from User import User
-
-# Load environment variables from .env file
-load_dotenv()
+from user import User
+from database import *
 
 app = Flask(__name__)
 app.secret_key = os.environ.get("APP_SECRET_KEY")
 login_manager = LoginManager()
 login_manager.init_app(app)
 
-# DB config used for accessing the database
-db_config = {
-    'host': os.environ.get("HOST"),
-    'user': os.environ.get("DATABASE_USERNAME"),
-    'password': os.environ.get("DATABASE_PASSWORD"),
-    'database': os.environ.get("DATABASE_NAME"),
-}
-
 @login_manager.user_loader
 def load_user(email):
 
-    # Establish a connection to the database
-    connection = mysql.connector.connect(**db_config)
-    cursor = connection.cursor()
-
-    # Retrieve the user with the given email
-    select_query = "SELECT * FROM users WHERE email = %s"
-    cursor.execute(select_query, (email,))
-    user = cursor.fetchone()
-
-    # Close the cursor and connection
-    cursor.close()
-    connection.close()
-
+    user = get_user_by_email(email)
+    
     if user:
         return User(user[2], *user[1:])
     return None
-
-def user_exists(email):
-    # Establish a connection to the database
-    connection = mysql.connector.connect(**db_config)
-    cursor = connection.cursor()
-
-    # Check if the user with the given email already exists
-    check_query = "SELECT * FROM users WHERE email = %s"
-    cursor.execute(check_query, (email,))
-    existing_user = cursor.fetchone()
-
-    # Close the cursor and connection
-    cursor.close()
-    connection.close()
-
-    return existing_user is not None
 
 def success_response(data=None, message=None):
     response = {'success': True}
@@ -72,46 +34,22 @@ def error_response(message, status_code):
 @app.route('/register', methods=['POST'])
 def register():
     user_data_json = request.get_json()
-    
-    # Establish a connection to the database
-    connection = mysql.connector.connect(**db_config)
-    cursor = connection.cursor()
 
     # Check if the user with the given email already exists
     if user_exists(user_data_json['email']):
         print(f"User with email {user_data_json['email']} already exists.")
+        return error_response('Email already taken', 401)
     else:
-        # Insert the new user into the database
-        insert_query = "INSERT INTO users (name, email, password) VALUES (%s, %s, %s)"
-        user_data = (user_data_json['name'], user_data_json['email'], user_data_json['password'])
-        cursor.execute(insert_query, user_data)
-        connection.commit()
+        add_user(user_data_json['name'], user_data_json['email'], user_data_json['password'])
         print(f"User {user_data_json['name']} registered successfully.")
-
-    # Close the cursor and connection
-    cursor.close()
-    connection.close()
 
     return success_response(message='Registration successful')
 
 # Route for handling login requests
 @app.route('/login', methods=['POST'])
 def login():
-    
     login_data_json = request.get_json()
-
-    # Establish a connection to the database
-    connection = mysql.connector.connect(**db_config)
-    cursor = connection.cursor()
-
-    # Retrieve the user with the given email
-    select_query = "SELECT * FROM users WHERE email = %s"
-    cursor.execute(select_query, (login_data_json['email'],))
-    user = cursor.fetchone()
-
-    # Close the cursor and connection
-    cursor.close()
-    connection.close()
+    user = get_user_by_email(login_data_json['email'])
 
     # Check if the user exists and the password is correct
     if user and user[3] == login_data_json['password']:
@@ -131,18 +69,7 @@ def edit_profile():
         print(f"User with email {user_data_json['email']} already exists.")
         return error_response('email is already in use', 401)
 
-    # Establish a connection to the database
-    connection = mysql.connector.connect(**db_config)
-    cursor = connection.cursor()
-
-    # Update the user's information
-    update_query = "UPDATE users SET name = %s, password = %s WHERE email = %s"
-    cursor.execute(update_query, (user_data_json['name'], user_data_json['password'], current_user.email))
-    connection.commit()
-
-    # Close the cursor and connection
-    cursor.close()
-    connection.close()
+    update_user(user_data_json['name'], user_data_json['password'], user_data_json['email'], current_user.email)
 
     return success_response(message='Profile updated successfully')
